TST: Reorganize the retrievals tests.
