+++ act/tests/retrievals/test_sonde.py
@@ -1,17 +1,6 @@
-' Unit tests for the ACT retrievals module. ' ''
-
-import glob
 import numpy as np
-import pytest
-import xarray as xr
+
 import act
-
-try:
-    import pysp2
-
-    PYSP2_AVAILABLE = True
-except ImportError:
-    PYSP2_AVAILABLE = False
 
 
 def test_get_stability_indices():
@@ -61,24 +50,6 @@ def test_get_stability_indices():
     sonde_ds.close()
 
 
-def test_generic_sobel_cbh():
-    ceil = act.io.arm.read_arm_netcdf(act.tests.sample_files.EXAMPLE_CEIL1)
-
-    ceil = ceil.resample(time='1min').nearest()
-    ceil = act.retrievals.cbh.generic_sobel_cbh(
-        ceil,
-        variable='backscatter',
-        height_dim='range',
-        var_thresh=1000.0,
-        fill_na=0.0,
-        edge_thresh=5,
-    )
-    cbh = ceil['cbh_sobel_backscatter'].values
-    assert cbh[500] == 615.0
-    assert cbh[1000] == 555.0
-    ceil.close()
-
-
 def test_calculate_precipitable_water():
     sonde_ds = act.io.arm.read_arm_netcdf(act.tests.sample_files.EXAMPLE_SONDE1)
     assert sonde_ds['tdry'].units == 'C', 'Temperature must be in Celsius'
@@ -91,92 +62,6 @@ def test_calculate_precipitable_water():
     sonde_ds.close()
 
 
-def test_doppler_lidar_winds():
-    # Process a single file
-    dl_ds = act.io.arm.read_arm_netcdf(act.tests.sample_files.EXAMPLE_DLPPI)
-    result = act.retrievals.doppler_lidar.compute_winds_from_ppi(dl_ds, intensity_name='intensity')
-    assert np.round(np.nansum(result['wind_speed'].values)).astype(int) == 1570
-    assert np.round(np.nansum(result['wind_direction'].values)).astype(int) == 32635
-    assert result['wind_speed'].attrs['units'] == 'm/s'
-    assert result['wind_direction'].attrs['units'] == 'degree'
-    assert result['height'].attrs['units'] == 'm'
-    dl_ds.close()
-
-    # Process multiple files
-    dl_ds = act.io.arm.read_arm_netcdf(act.tests.sample_files.EXAMPLE_DLPPI_MULTI)
-    del dl_ds['range'].attrs['units']
-    result = act.retrievals.doppler_lidar.compute_winds_from_ppi(dl_ds)
-    assert np.round(np.nansum(result['wind_speed'].values)).astype(int) == 2854
-    assert np.round(np.nansum(result['wind_direction'].values)).astype(int) == 64986
-    dl_ds.close()
-
-
-def test_aeri2irt():
-    aeri_ds = act.io.arm.read_arm_netcdf(act.tests.sample_files.EXAMPLE_AERI)
-    aeri_ds = act.retrievals.aeri.aeri2irt(aeri_ds)
-    assert np.round(np.nansum(aeri_ds['aeri_irt_equiv_temperature'].values)).astype(int) == 17372
-    np.testing.assert_almost_equal(
-        aeri_ds['aeri_irt_equiv_temperature'].values[7], 286.081, decimal=3
-    )
-    np.testing.assert_almost_equal(
-        aeri_ds['aeri_irt_equiv_temperature'].values[-10], 285.366, decimal=3
-    )
-    aeri_ds.close()
-    del aeri_ds
-
-
-def test_sst():
-    ds = act.io.arm.read_arm_netcdf(act.tests.sample_files.EXAMPLE_IRTSST)
-    ds = act.retrievals.irt.sst_from_irt(ds)
-    np.testing.assert_almost_equal(ds['sea_surface_temperature'].values[0], 278.901, decimal=3)
-    np.testing.assert_almost_equal(ds['sea_surface_temperature'].values[-1], 279.291, decimal=3)
-    assert np.round(np.nansum(ds['sea_surface_temperature'].values)).astype(int) == 6699
-    ds.close()
-
-
-def test_calculate_sirs_variable():
-    sirs_ds = act.io.arm.read_arm_netcdf(act.tests.sample_files.EXAMPLE_SIRS)
-    met_ds = act.io.arm.read_arm_netcdf(act.tests.sample_files.EXAMPLE_MET1)
-
-    ds = act.retrievals.radiation.calculate_dsh_from_dsdh_sdn(sirs_ds)
-    assert np.isclose(np.nansum(ds['derived_down_short_hemisp'].values), 61157.71, atol=0.1)
-
-    ds = act.retrievals.radiation.calculate_irradiance_stats(
-        ds,
-        variable='derived_down_short_hemisp',
-        variable2='down_short_hemisp',
-        threshold=60,
-    )
-
-    assert np.isclose(np.nansum(ds['diff_derived_down_short_hemisp'].values), 1335.12, atol=0.1)
-    assert np.isclose(np.nansum(ds['ratio_derived_down_short_hemisp'].values), 400.31, atol=0.1)
-
-    ds = act.retrievals.radiation.calculate_net_radiation(ds, smooth=30)
-    assert np.ceil(np.nansum(ds['net_radiation'].values)) == 21915
-    assert np.ceil(np.nansum(ds['net_radiation_smoothed'].values)) == 22316
-
-    ds = act.retrievals.radiation.calculate_longwave_radiation(
-        ds,
-        temperature_var='temp_mean',
-        vapor_pressure_var='vapor_pressure_mean',
-        met_ds=met_ds,
-    )
-    assert np.ceil(ds['monteith_clear'].values[25]) == 239
-    assert np.ceil(ds['monteith_cloudy'].values[30]) == 318
-    assert np.ceil(ds['prata_clear'].values[35]) == 234
-
-    new_ds = xr.merge([sirs_ds, met_ds], compat='override')
-    ds = act.retrievals.radiation.calculate_longwave_radiation(
-        new_ds, temperature_var='temp_mean', vapor_pressure_var='vapor_pressure_mean'
-    )
-    assert np.ceil(ds['monteith_clear'].values[25]) == 239
-    assert np.ceil(ds['monteith_cloudy'].values[30]) == 318
-    assert np.ceil(ds['prata_clear'].values[35]) == 234
-
-    sirs_ds.close()
-    met_ds.close()
-
-
 def test_calculate_pbl_liu_liang():
     files = act.tests.sample_files.EXAMPLE_TWP_SONDE_20060121.copy()
     files2 = act.tests.sample_files.EXAMPLE_SONDE1
@@ -185,8 +70,8 @@ def test_calculate_pbl_liu_liang():
 
     pblht = []
     pbl_regime = []
-    for i, r in enumerate(files):
-        ds = act.io.arm.read_arm_netcdf(r)
+    for file in files:
+        ds = act.io.arm.read_arm_netcdf(file)
         ds['tdry'].attrs['units'] = 'degree_Celsius'
         ds = act.retrievals.sonde.calculate_pbl_liu_liang(ds, smooth_height=10)
         pblht.append(float(ds['pblht_liu_liang'].values))
@@ -247,26 +132,3 @@ def test_calculate_heffter_pbl():
     np.testing.assert_almost_equal(ds['potential_temperature_ss'].values[4], 298.4, 1)
     assert np.sum(ds['bottom_inversion'].values) == 7426
     assert np.sum(ds['top_inversion'].values) == 7903
-
-
-@pytest.mark.skipif(not PYSP2_AVAILABLE, reason='PySP2 is not installed.')
-def test_sp2_waveform_stats():
-    my_sp2b = act.io.read_sp2(act.tests.EXAMPLE_SP2B)
-    my_ini = act.tests.EXAMPLE_INI
-    my_binary = act.qc.get_waveform_statistics(my_sp2b, my_ini, parallel=False)
-    assert my_binary.PkHt_ch1.max() == 62669.4
-    np.testing.assert_almost_equal(np.nanmax(my_binary.PkHt_ch0.values), 98708.92915295, decimal=1)
-    np.testing.assert_almost_equal(np.nanmax(my_binary.PkHt_ch4.values), 65088.39598033, decimal=1)
-
-
-@pytest.mark.skipif(not PYSP2_AVAILABLE, reason='PySP2 is not installed.')
-def test_sp2_psds():
-    my_sp2b = act.io.read_sp2(act.tests.EXAMPLE_SP2B)
-    my_ini = act.tests.EXAMPLE_INI
-    my_binary = act.qc.get_waveform_statistics(my_sp2b, my_ini, parallel=False)
-    my_hk = act.io.read_hk_file(act.tests.EXAMPLE_HK)
-    my_binary = act.retrievals.calc_sp2_diams_masses(my_binary)
-    ScatRejectKey = my_binary['ScatRejectKey'].values
-    assert np.nanmax(my_binary['ScatDiaBC50'].values[ScatRejectKey == 0]) < 1000.0
-    my_psds = act.retrievals.process_sp2_psds(my_binary, my_hk, my_ini)
-    np.testing.assert_almost_equal(my_psds['NumConcIncan'].max(), 0.95805343)
